Move calculations into evaluators
